refactor: extract ContentView+Toolbar.swift (check_shell_chrome, 2/2)
ContentView.swift was 1022 lines, over the 1000-line shell limit. Extracted the
toolbar — a cohesive slice with an obvious name: everything that builds the window
toolbar, plus the toolbar-item identity vocabulary only the toolbar speaks.

- New: Views/ContentView+Toolbar.swift (247 lines) — ContentToolbarID + the whole
  `// MARK: - Toolbar Content` extension (mainToolbarContent, leading/trailing/
  content-pane/principal sections, view-mode menu, inspector toggle). Beside
  ContentView+ViewBuilders.swift; a stranger finds the toolbar by filename.
- ContentView.swift: 1022 → 785. Both shell files now under the limit.

Visibility widenings — the only semantic-adjacent edits, every one FORCED by
`private` being FILE-scoped once the code moves file, and every one verified as
actually needed (three members nothing calls across the boundary stay private):
- `private enum ContentToolbarID` → internal: the body still names these ids.
- `@State private var focusedDocument` → internal: the toolbar's selection sync.
- contentPaneToolbarContent / trailingToolbarContent / principalToolbarContent /
  inspectorToggleButton / syncFocusedDocumentSelection → internal: ContentView's
  body composes its own .toolbar from these sections and calls them.
- leadingToolbarContent / viewDisplayModeMenu / platformViewMenuButton stay
  private — nothing outside the file touches them.

The moved block is otherwise byte-identical to HEAD (verified by diff): same items,
same placements, same order, no new bodies or computed properties, so the
type-checker's job only shrinks.

BUILD SUCCEEDED (xcodebuild, own SYMROOT, compile only). The first attempt FAILED
with five 'inaccessible due to private protection level' errors — I had checked
that the toolbar didn't need the struct's privates, but not the reverse, that the
body calls the extension's privates. Fixed and rebuilt green.

project.pbxproj: 4 entries; plutil -lint passes. swiftlint clean on both files.

check_shell_chrome's OVERSIZE rule now PASSES (1320→757 and 1022→785). The script
still exits 1 on its separate §6b rule (SettingsView's `= FeatureManager.shared`)
— the @Observable FeatureManager migration, next commit, kept apart to bisect.

(cherry picked from commit 08f6feac8d96d0288160aae221e4572749ad139d)
